Listener/requirements.txt
requests==2.25.1
pycryptodome==3.9.9
